+++ src/main/java/MMS/inventory/services/Impl/PatientServiceImpl.java
@@ -2,6 +2,7 @@ package MMS.inventory.services.Impl;
 
 import MMS.inventory.DTO.PatientDto;
 import MMS.inventory.DTO.mapper.PatientMapper;
+import MMS.inventory.Exception.ResourceNotFoundException;
 import MMS.inventory.model.Patient;
 import MMS.inventory.repository.PatientRepository;
 import MMS.inventory.services.PatientService;
@@ -20,23 +21,31 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public PatientDto getPatient(Long patientId) {
-        return patientMapper.toPatientDto(patientRepository.findById(patientId).get());
+        Patient patient = patientRepository.findById(patientId).orElseThrow(() -> new ResourceNotFoundException("Patient not found with id: " + patientId));
+        return patientMapper.toPatientDto(patient);
     }
 
     @Override
     public PatientDto getPatientByEmail(String email) {
-        return patientMapper.toPatientDto(patientRepository.findByEmail(email));
+       Patient patient = patientRepository.findByEmail(email);
+         if(patient == null) {
+              throw new ResourceNotFoundException("Patient not found with email: " + email);
+         }
+        return patientMapper.toPatientDto(patient);
     }
 
     @Override
     public PatientDto createPatient(PatientDto patient) {
+        if(patient == null) {
+            throw new ResourceNotFoundException("Patient cannot be null");
+        }
         Patient patientToSave = patientMapper.toPatient(patient);
         return patientMapper.toPatientDto(patientRepository.save(patientToSave));
     }
 
     @Override
     public PatientDto updatePatientById(Long patientId, PatientDto patientDto) {
-        Patient patientToUpdate = patientRepository.findById(patientId).get();
+        Patient patientToUpdate = patientRepository.findById(patientId).orElseThrow(() -> new ResourceNotFoundException("Patient not found with id: " + patientId));
         PatientDto patient = patientMapper.toPatientDto(patientToUpdate);
         patientToUpdate.getGeneralDetail().setName(patient.getPatientName());
         patientToUpdate.getGeneralDetail().setGender(patient.getPatientGender());
@@ -48,7 +57,7 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public PatientDto patchPatientById(Long patientId, PatientDto patient) {
-        Patient patientToUpdate = patientRepository.findById(patientId).get();
+        Patient patientToUpdate = patientRepository.findById(patientId).orElseThrow(() -> new ResourceNotFoundException("Patient not found with id: " + patientId));
         patientToUpdate.getGeneralDetail().setName(patient.getPatientName());
         patientToUpdate.getGeneralDetail().setGender(patient.getPatientGender());
         patientToUpdate.getGeneralDetail().setEmail(patient.getPatientEmail());
@@ -59,6 +68,9 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public void deletePatientById(Long patientId) {
+        if(!patientRepository.existsById(patientId)) {
+            throw new ResourceNotFoundException("Patient not found with id: " + patientId);
+        }
         patientRepository.deleteById(patientId);
     }
